+++ src/main/java/mod/gottsch/forge/treasure2/core/world/feature/gen/TreasureFeatureGenerators.java
@@ -19,12 +19,16 @@ package mod.gottsch.forge.treasure2.core.world.feature.gen;
 
 import java.util.Optional;
 
+import mod.gottsch.forge.treasure2.Treasure;
+import mod.gottsch.forge.treasure2.core.block.DeferredGeneratorBlock;
+import mod.gottsch.forge.treasure2.core.block.TreasureBlocks;
 import mod.gottsch.forge.treasure2.core.config.ChestFeaturesConfiguration.Generator;
 import mod.gottsch.forge.treasure2.core.config.Config;
 import mod.gottsch.forge.treasure2.core.registry.FeatureGeneratorRegistry;
 import mod.gottsch.forge.treasure2.core.util.ModUtil;
 import mod.gottsch.forge.treasure2.core.world.feature.FeatureType;
 import mod.gottsch.forge.treasure2.core.world.feature.gen.selector.*;
+import net.minecraft.resources.ResourceLocation;
 
 /**
  * 
@@ -35,21 +39,27 @@ public class TreasureFeatureGenerators {
 
 	// feature generators
 	public static final IFeatureGenerator SIMPLE_SURFACE_FEATURE_GENERATOR = new SimpleSurfaceChestFeatureGenerator();
-	public static final IFeatureGenerator PIT_FEATURE_GENERATOR = new PitChestFeatureGenerator();
 
 	// the actual feature generators that are called by a Tickable Block Entity
 	public static final IFeatureGenerator WITHER_FEATURE_GENERATOR = new WitherFeatureGenerator();
 	public static final IFeatureGenerator SURFACE_STRUCTURE_FEATURE_GENERATOR = new SurfaceStructureFeatureGenerator();
 	public static final IFeatureGenerator SUBAQUATIC_FEATURE_GENERATOR = new SubaquaticStructureFeatureGenerator();
+	public static final IFeatureGenerator PIT_FEATURE_GENERATOR = new PitChestFeatureGenerator();
 
 	/*
 	 * a deferred feature generators that places a Tickable Block Entity, which calls the actual feature generator to generate at a specific player proximity.
 	 */
-	public static final IFeatureGenerator DEFERRED_WITHER_FEATURE_GENERATOR = new DeferredWitherFeatureGenerator();
-	public static final IFeatureGenerator DEFERRED_SURFACE_STRUCTURE_FEATURE_GENERATOR = new DeferredSurfaceStructureFeatureGenerator();
-	public static final IFeatureGenerator DEFERRED_SUBAQUATIC_FEATURE_GENERATOR = new DeferredSubaquaticStructureFeatureGenerator();
+	public static final IFeatureGenerator DEFERRED_WITHER_FEATURE_GENERATOR =
+			new DeferredFeatureGenerator(new ResourceLocation(Treasure.MODID, "deferred_wither"), (DeferredGeneratorBlock) TreasureBlocks.DEFERRED_WITHER_TREE_GENERATOR.get());
+	public static final IFeatureGenerator DEFERRED_SURFACE_STRUCTURE_FEATURE_GENERATOR =
+			new DeferredFeatureGenerator(new ResourceLocation(Treasure.MODID, "deferred_subaquatic_structure"), (DeferredGeneratorBlock) TreasureBlocks.DEFERRED_SUBAQUATIC_GENERATOR.get());
+	public static final IFeatureGenerator DEFERRED_SUBAQUATIC_FEATURE_GENERATOR =
+			new DeferredFeatureGenerator(new ResourceLocation(Treasure.MODID, "deferred_surface_structure"), (DeferredGeneratorBlock) TreasureBlocks.DEFERRED_SURFACE_GENERATOR.get());
 
-	
+	//	public static final IFeatureGenerator DEFERRED_PIT_FEATURE_GENERATOR = new DeferredPitFeatureGenerator();
+	public static final IFeatureGenerator DEFERRED_PIT_FEATURE_GENERATOR =
+		new DeferredFeatureGenerator(new ResourceLocation(Treasure.MODID, "deferred_pit"), (DeferredGeneratorBlock) TreasureBlocks.DEFERRED_PIT_GENERATOR.get());
+
 	// feature generator selectors
 	public static final IFeatureGeneratorSelector STANDARD_CHEST_FEATURE_GENERATOR_SELECTOR = new WeightedChestFeatureGeneratorSelector();
 	public static final IFeatureGeneratorSelector WITHER_FEATURE_GENERATOR_SELECTOR = new DeferredWitherFeatureGeneratorSelector();
@@ -66,7 +76,7 @@ public class TreasureFeatureGenerators {
 		// setup pre-made feature generators
 		WeightedChestFeatureGeneratorSelector selector = (WeightedChestFeatureGeneratorSelector)STANDARD_CHEST_FEATURE_GENERATOR_SELECTOR;
 		selector.add(10, TreasureFeatureGenerators.SIMPLE_SURFACE_FEATURE_GENERATOR);
-		selector.add(65, TreasureFeatureGenerators.PIT_FEATURE_GENERATOR);
+		selector.add(65, TreasureFeatureGenerators.DEFERRED_PIT_FEATURE_GENERATOR);
 		selector.add(25, DEFERRED_SURFACE_STRUCTURE_FEATURE_GENERATOR);
 	}