+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.8.4] - 2024-07-23
+## [3.9.0] - 2024-07-28
+
+### Changed
+- Fixed Deferred generator block entities using the wrong config-based conditional statement.
+
+### Added
+- Added Deferred generators for Surface, Subaquatic structures, and Pits.
+
+
+## [3.8.5] - 2024-07-23
 
 ### Changed
 - Fixed pit shaft and pit room alignments! For real this time.